feat: support changing canvas margin (#37)
add new input property "margin" to the component, and "qrCodeMargin" to
the directive that allows configuring the amount of whitespace is
surrounding the image that is output to the canvas.

Keep backwards compatibility by defaulting to 4, however I suspect this
is probably easiest to reason about if it was 0 and normal css padding/margin
was used in it's place, so may change the default in the next major
version.

closes #36